+++ core/data/src/main/kotlin/com/manuelnunez/apps/core/data/repository/SeeMoreRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.manuelnunez.apps.core.data.repository
 
 import com.manuelnunez.apps.core.data.datasource.PexelsCatsRemoteDataSource
-import com.manuelnunez.apps.feature.seemore.domain.repository.SeeMoreRepository
+import com.manuelnunez.apps.features.seemore.domain.repository.SeeMoreRepository
 import javax.inject.Inject
 
 class SeeMoreRepositoryImpl
