+++ features/detail/ui/src/main/kotlin/com/manuelnunez/apps/features/detail/ui/navigation/DetailNavigation.kt
@@ -1,0 +1,26 @@
+package com.manuelnunez.apps.features.detail.ui.navigation
+
+import android.os.Bundle
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavOptionsBuilder
+import androidx.navigation.compose.composable
+import androidx.navigation.navOptions
+import com.manuelnunez.apps.core.common.navigation.navigate
+import com.manuelnunez.apps.core.domain.model.Item
+import com.manuelnunez.apps.features.detail.ui.DetailView
+
+const val DETAIL_ITEM = "myItem"
+const val DETAIL_ROUTE = "detail"
+
+fun NavController.navigateToDetail(item: Item, navOptions: NavOptionsBuilder.() -> Unit = {}) {
+  val bundle = Bundle().apply { putParcelable(DETAIL_ITEM, item) }
+  navigate(route = DETAIL_ROUTE, args = bundle, navOptions = navOptions(navOptions))
+}
+
+fun NavGraphBuilder.detailScreen(onBackClick: () -> Unit) {
+  composable(DETAIL_ROUTE) { backStackEntry ->
+    val myItem = backStackEntry.arguments?.getParcelable<Item>(DETAIL_ITEM)
+    DetailView(onBackClick = onBackClick, item = myItem)
+  }
+}
