--- features/seemore/domain/src/main/kotlin/com/manuelnunez/apps/feature/seemore/domain/repository/SeeMoreRepository.kt
@@ -1,9 +1,0 @@
-package com.manuelnunez.apps.feature.seemore.domain.repository
-
-import androidx.paging.PagingData
-import com.manuelnunez.apps.core.domain.model.Item
-import kotlinx.coroutines.flow.Flow
-
-interface SeeMoreRepository {
-  fun getAllItems(): Flow<PagingData<Item>>
-}
